modify create, edit todo logic

+++ src/main/java/com/example/weup/service/TodoService.java
@@ -48,24 +48,24 @@ public class TodoService {
 
         Todo todo = Todo.builder()
                 .project(project)
-                .todoName(createTodoRequestDTO.getTodoName())
-                .startDate(createTodoRequestDTO.getStartDate())
-                .endDate(createTodoRequestDTO.getEndDate())
+//                .todoName(createTodoRequestDTO.getTodoName())
+//                .startDate(createTodoRequestDTO.getStartDate())
+//                .endDate(createTodoRequestDTO.getEndDate())
                 .build();
 
         todoRepository.save(todo);
 
-        for (Long memberId : createTodoRequestDTO.getMemberIds()) {
-            Member member = memberRepository.findById(memberId)
-                    .orElseThrow(() -> new GeneralException(ErrorInfo.USER_NOT_FOUND));
-
-            TodoMember todoMember = TodoMember.builder()
-                    .todo(todo)
-                    .member(member)
-                    .build();
-
-            todoMemberRepository.save(todoMember);
-        }
+//        for (Long memberId : createTodoRequestDTO.getMemberIds()) {
+//            Member member = memberRepository.findById(memberId)
+//                    .orElseThrow(() -> new GeneralException(ErrorInfo.USER_NOT_FOUND));
+//
+//            TodoMember todoMember = TodoMember.builder()
+//                    .todo(todo)
+//                    .member(member)
+//                    .build();
+//
+//            todoMemberRepository.save(todoMember);
+//        }
     }
 
     public List<TodoListResponseDTO> getTodoList(Long userId, Long projectId) {
@@ -132,24 +132,40 @@ public class TodoService {
             throw new GeneralException(ErrorInfo.FORBIDDEN);
         }
 
-        todo.setTodoName(editTodoRequestDTO.getTodoName());
-        todo.setStartDate(editTodoRequestDTO.getStartDate());
-        todo.setEndDate(editTodoRequestDTO.getEndDate());
+        if (editTodoRequestDTO.getTodoName() != null) {
+            todo.setTodoName(editTodoRequestDTO.getTodoName());
+        }
 
-        todoMemberRepository.deleteByTodo_TodoId(editTodoRequestDTO.getTodoId());
+        if ((editTodoRequestDTO.getStartDate() != null) && (editTodoRequestDTO.getEndDate() != null)) {
+            todo.setStartDate(editTodoRequestDTO.getStartDate());
+            todo.setEndDate(editTodoRequestDTO.getEndDate());
+        }
 
-        editTodoRequestDTO.getMemberIds().stream()
-                .map(memberId -> memberRepository.findById(memberId)
-                        .orElseThrow(() -> new GeneralException(ErrorInfo.USER_NOT_FOUND)))
-                .map(member -> TodoMember.builder()
-                        .todo(todo)
-                        .member(member)
-                        .build())
-                .forEach(todoMemberRepository::save);
+        todoRepository.save(todo);
+
+        log.error(todo.getTodoName());
+
+        if (editTodoRequestDTO.getMemberIds() != null) {
+            todoMemberRepository.deleteByTodo_TodoId(editTodoRequestDTO.getTodoId());
+
+            editTodoRequestDTO.getMemberIds().stream()
+                    .map(memberId -> memberRepository.findById(memberId)
+                            .orElseThrow(() -> new GeneralException(ErrorInfo.USER_NOT_FOUND)))
+                    .map(member -> TodoMember.builder()
+                            .todo(todo)
+                            .member(member)
+                            .build())
+                    .forEach(todoMemberRepository::save);
+        }
     }
+
 
     @Transactional
     public void editTodoStatus(Long userId, EditTodoStatusRequestDTO editTodoStatusRequestDTO) {
+
+        log.debug("edit todo status service in");
+        log.debug("todo id : {}", editTodoStatusRequestDTO.getTodoId());
+        log.debug("status : {}", editTodoStatusRequestDTO.getStatus());
 
         Todo todo = todoRepository.findById(editTodoStatusRequestDTO.getTodoId())
                 .orElseThrow(() -> new GeneralException(ErrorInfo.TODO_NOT_FOUND));
